Fixed python_options.h header - a macro error was causing some segfaults. Cleaned up related sizing issues across ssystem.cc and energymodel.h. Need to work on the newObject ref counting.

+++ include/ssystem.h
@@ -71,8 +71,8 @@ class SimulationSystem
   bool boltzmann_sampling;
   bool use_fixed_random_seed;
   bool initial_trajectory;
-  int simulation_mode;
-  int simulation_count_remaining;
+  long simulation_mode;
+  long simulation_count_remaining;
 };
 
 #endif
